:recycle: Resolve cycle
