+++ graphix/command.py
@@ -10,7 +10,7 @@ import numpy as np
 from pydantic import BaseModel
 
 import graphix.clifford
-from graphix.pauli import Plane
+from graphix.pauli import Pauli, Plane
 
 if TYPE_CHECKING:
     from graphix.clifford import Clifford
@@ -72,7 +72,7 @@ class M(Command):
                 pass
             else:
                 raise RuntimeError(f"{gate} should be either I, H, S or Z.")
-        update = graphix.pauli.MeasureUpdate.compute(self.plane, False, False, clifford)
+        update = MeasureUpdate.compute(self.plane, False, False, clifford)
         return M(
             node=self.node,
             plane=update.new_plane,
@@ -143,3 +143,32 @@ class T(Command):
     """
 
     kind: CommandKind = CommandKind.T
+
+
+class MeasureUpdate(BaseModel):
+    new_plane: Plane
+    coeff: int
+    add_term: float
+
+    @staticmethod
+    def compute(plane: Plane, s: bool, t: bool, clifford: graphix.clifford.Clifford) -> MeasureUpdate:
+        gates = list(map(Pauli.from_axis, plane.axes))
+        if s:
+            clifford = graphix.clifford.X @ clifford
+        if t:
+            clifford = graphix.clifford.Z @ clifford
+        gates = list(map(clifford.measure, gates))
+        new_plane = Plane.from_axes(*(gate.axis for gate in gates))
+        cos_pauli = clifford.measure(Pauli.from_axis(plane.cos))
+        sin_pauli = clifford.measure(Pauli.from_axis(plane.sin))
+        exchange = cos_pauli.axis != new_plane.cos
+        if exchange == (cos_pauli.unit.sign == sin_pauli.unit.sign):
+            coeff = -1
+        else:
+            coeff = 1
+        add_term = 0
+        if cos_pauli.unit.sign:
+            add_term += np.pi
+        if exchange:
+            add_term = np.pi / 2 - add_term
+        return MeasureUpdate(new_plane=new_plane, coeff=coeff, add_term=add_term)
